+++ gradsflow/__init__.py
@@ -1,7 +1,8 @@
 """An AutoML Library made with Optuna and PyTorch Lightning"""
 
-from gradsflow.autoclassifier.image import AutoImageClassifier
-from gradsflow.autoclassifier.text import AutoTextClassifier
-from gradsflow.automodel import AutoModel
+from gradsflow.tasks.autoclassification.text import AutoTextClassifier
+from gradsflow.tasks.autoclassification.image import AutoImageClassifier
+from gradsflow.tasks.summarization import AutoSummarization
+from gradsflow.core.automodel import AutoModel
 
 __version__ = "0.0.1"
