--- gradsflow/automodel/__init__.py
@@ -1,1 +1,0 @@
-from .automodel import AutoModel
